backend/agent/requirements.txt
# LangGraph and Core
langgraph>=0.2.0
langgraph-checkpoint-mongodb>=0.1.0
langchain>=0.3.0
langchain-core>=0.3.0

# LLM Providers
langchain-anthropic>=0.3.0

# MongoDB
langchain-mongodb>=0.2.0
pymongo>=4.6.0
motor>=3.3.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment
python-dotenv>=1.0.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
sse-starlette>=1.8.0

# Async
aiohttp>=3.9.0
